Chore : 프로필 폴더 합치고 라우트 수정
